// src/app/SistemaWarhammerEspadaDaPerdicao/InterfaceWarhammer/WarhammerInterface.interface.ts
export interface IWarhammer{
    nome: string;
    nivelMortalidade: number;
    nivelPoder: number;
    armas: string;
    armadura: string;
    corArmadura: string;
    nivelDeLealdade: number;
    nivelDeCoragem: number;
    nivelDeInteligencia: number;
    nivelDeForca: number;
    nivelDeAgilidade: number;
    nivelDeHabilidade: number;
}